Fixed bug in updates for challenge total.  Removed inadvertant commenting.
The challenge total was supposed to be updated when the calculated
total point was not equal to the stored point total.  Instead it was
being updated when the student score was not equal to the stored point
total for the challenge.  This both meant unnecessary database updates
and that when the challenge was taken the very first time, if the first
person to take it bombed it, it would not get updated.

Also, someone had commented out the code which was checking to make
sure that the same attempt was not submitted twice.  I'm pretty
sure this was inadvertant, so I uncommented it.